Reformat eve_ui and eve_ui_keyboard libraries
