+++ lib/eve_ui/source/eve_ui_images.c
@@ -120,43 +120,3 @@ uint32_t eve_ui_load_image(const uint8_t *img_data, uint32_t start_addr, uint8_t
 	return start_addr;
 }
 
-extern const uint8_t img_bridgetek_logo_jpg[] asm("img_bridgetek_logo_jpg");
-extern const uint8_t img_settings_jpg[] asm("img_settings_jpg");
-extern const uint8_t img_cancel_jpg[] asm("img_cancel_jpg");
-extern const uint8_t img_tick_jpg[] asm("img_z_jpg");
-extern const uint8_t img_refresh_jpg[] asm("img_refresh_jpg");
-extern const uint8_t img_keypad_jpg[] asm("img_keypad_jpg");
-extern const uint8_t img_keyboard_jpg[] asm("img_keyboard_jpg");
-extern const uint8_t img_media_jpg[] asm("img_media_jpg");
-
-uint32_t eve_ui_load_images(uint32_t start_addr)
-{
-	uint32_t dummy = start_addr;
-
-	// Load images statically and sequentially.
-	dummy = eve_ui_load_image(img_bridgetek_logo_jpg, dummy,
-			BITMAP_BRIDGETEK_LOGO, &img_bridgetek_logo_width, &img_bridgetek_logo_height);
-
-	dummy = eve_ui_load_image(img_settings_jpg, dummy,
-			BITMAP_SETTINGS, &img_settings_width, &img_settings_height);
-
-	dummy = eve_ui_load_image(img_cancel_jpg, dummy,
-			BITMAP_CANCEL, &img_cancel_width, &img_cancel_height);
-
-	dummy = eve_ui_load_image(img_tick_jpg, dummy,
-			BITMAP_SAVE, &img_tick_width, &img_tick_height);
-
-	dummy = eve_ui_load_image(img_refresh_jpg, dummy,
-			BITMAP_REFRESH, &img_refresh_width, &img_refresh_height);
-
-	dummy = eve_ui_load_image(img_keypad_jpg, dummy,
-			BITMAP_KEYPAD, &img_keypad_width, &img_keypad_height);
-
-	dummy = eve_ui_load_image(img_keyboard_jpg, dummy,
-			BITMAP_KEYBOARD, &img_keyboard_width, &img_keyboard_height);
-
-	dummy = eve_ui_load_image(img_media_jpg, dummy,
-			BITMAP_MEDIA, &img_media_width, &img_media_height);
-
-	return dummy;
-}
