+++ lib/eve_ui_keyboard/Images/S/bridgetek_logo.s
