+++ lib/eve_ui_keyboard/Images/S/cancel.s
