+++ lib/eve_ui_keyboard/Images/S/settings.s
