+++ lib/eve_ui_keyboard/Images/S/tick.s
